Move BoardInitialPosition class to separate module
